% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion2D.R
\name{.calcSegment}
\alias{.calcSegment}
\title{Calculate Segment Coordinates}
\usage{
.calcSegment(wall_data, direction, x_dist, y_dist)
}
\arguments{
\item{wall_data}{Data frame. Wall position data}

\item{direction}{Character. Wall direction}

\item{x_dist}{Numeric. X-axis distance scale}

\item{y_dist}{Numeric. Y-axis distance scale}
}
\value{
List of segment coordinates
}
\description{
Calculates start and end coordinates for wall segments
}
\keyword{internal}
