% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.countDistance}
\alias{.countDistance}
\title{Normalize Coordinate Matrix by Feature Distance}
\usage{
.countDistance(dist_input)
}
\arguments{
\item{dist_input}{Matrix or data.frame. Contains spatial coordinates.}
}
\value{
A list containing the normalized matrix and distance values for
each dimension
}
\description{
Normalizes coordinates by dividing each dimension by its
minimum distance
}
\keyword{internal}
