% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion2D.R
\name{.createSegmentLayer}
\alias{.createSegmentLayer}
\title{Create Segment Layer for Wall Visualization}
\usage{
.createSegmentLayer(wall_data, direction, color, x_dist, y_dist)
}
\arguments{
\item{wall_data}{Data frame. Wall coordinates}

\item{direction}{Character. Wall direction}

\item{color}{Character. Wall color}

\item{x_dist}{Numeric. X-axis distance scale}

\item{y_dist}{Numeric. Y-axis distance scale}
}
\value{
A ggplot2 layer object or NULL
}
\description{
Creates a ggplot2 layer for wall segments
}
\keyword{internal}
