% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion2D.R
\name{.findNeighbor}
\alias{.findNeighbor}
\title{Find Neighboring Features}
\usage{
.findNeighbor(self, others, type = c("right", "left", "top", "bottom"))
}
\arguments{
\item{self}{Numeric vector. Reference point coordinates}

\item{others}{Matrix. Coordinates of other features}

\item{type}{Character. Direction to check ("right", "left", "top", "bottom")}
}
\value{
Logical vector indicating neighbor presence
}
\description{
Identifies neighboring features in specified directions
}
\keyword{internal}
