% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion2D.R
\name{.pvalAnnotation}
\alias{.pvalAnnotation}
\title{Create P-value Annotation}
\usage{
.pvalAnnotation(pval)
}
\arguments{
\item{pval}{Numeric. P-value to convert}
}
\value{
Character string with significance symbols
}
\description{
Converts p-values to significance symbols
}
\keyword{internal}
