% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeLipidRegion.R
\name{.resultSummary}
\alias{.resultSummary}
\title{Summarize Results}
\usage{
.resultSummary(
  X,
  X.info,
  group,
  smooth.stat,
  smooth.stat.permute,
  region.stat.obs,
  dimension,
  permute_time,
  ...
)
}
\arguments{
\item{X}{Matrix. Lipid abundance data matrix.}

\item{X.info}{Data frame. Lipid feature information.}

\item{group}{Vector. Group assignments.}

\item{smooth.stat}{Numeric vector. Smoothed statistics.}

\item{smooth.stat.permute}{Matrix. Permutation results.}

\item{region.stat.obs}{Numeric vector. Observed region statistics.}

\item{dimension}{Integer. Dimension of the analysis.}

\item{permute_time}{Integer. Number of permutations performed.}

\item{...}{Additional arguments.}
}
\value{
Data frame containing formatted results.
}
\description{
Summarize and format lipid trend analysis results
}
\keyword{internal}
