% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{.split_chain}
\alias{.split_chain}
\title{Get Split Chain Status from LipidTrendSE}
\usage{
.split_chain(object)
}
\arguments{
\item{object}{A LipidTrendSE object}
}
\value{
Logical indicating if analysis was split by chain
}
\description{
Get Split Chain Status from LipidTrendSE
}
\keyword{internal}
