% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion1D.R
\name{.trendPlot1D}
\alias{.trendPlot1D}
\title{Create 1D result plot}
\usage{
.trendPlot1D(res, p_cutoff, y_scale)
}
\arguments{
\item{res}{A data.frame containing columns for coordinates, abundance values,
direction, and adjusted p-values}

\item{p_cutoff}{Numeric value for significance cutoff.
Default is \code{0.05}}

\item{y_scale}{Character. Choose one of the y-axis scales: "identity",
"log2", "log10", or "sqrt". Default is \code{"identity"}}
}
\value{
A ggplot2 object
}
\description{
Creates a ggplot visualization showing trends between case and
control groups with highlighted significant regions.
}
\keyword{internal}
