% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getEvenChainResult}
\alias{getEvenChainResult}
\alias{getEvenChainResult,LipidTrendSE-method}
\title{Get Even Chain Result from LipidTrendSE}
\usage{
getEvenChainResult(object)

\S4method{getEvenChainResult}{LipidTrendSE}(object)
}
\arguments{
\item{object}{A LipidTrendSE object}
}
\value{
A data frame containing even chain result or NULL
}
\description{
Get Even Chain Result from LipidTrendSE
}
\examples{
data("lipid_se_CL")
sub <- lipid_se_CL[seq_len(10), ]
res_se <- analyzeLipidRegion(
    lipid_se=sub, ref_group="sgCtrl", split_chain=TRUE,
    chain_col="chain", radius=3, own_contri=0.5, permute_time=100)
# Get complete result summary
results <- getEvenChainResult(res_se)
}
