% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getOddChainResult}
\alias{getOddChainResult}
\alias{getOddChainResult,LipidTrendSE-method}
\title{Get Odd Chain Result from LipidTrendSE}
\usage{
getOddChainResult(object)

\S4method{getOddChainResult}{LipidTrendSE}(object)
}
\arguments{
\item{object}{A LipidTrendSE object}
}
\value{
A data frame containing odd chain result or NULL
}
\description{
Get Odd Chain Result from LipidTrendSE
}
\examples{
data("lipid_se_CL")
res_se <- analyzeLipidRegion(
    lipid_se=lipid_se_CL, ref_group="sgCtrl", split_chain=TRUE,
    chain_col="chain", radius=3, own_contri=0.5, permute_time=100)
# Get complete result summary
results <- getOddChainResult(res_se)
}
