% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getResult}
\alias{getResult}
\alias{getResult,LipidTrendSE-method}
\title{Get Result from LipidTrendSE}
\usage{
getResult(object)

\S4method{getResult}{LipidTrendSE}(object)
}
\arguments{
\item{object}{A LipidTrendSE object}
}
\value{
A data frame containing analysis results or NULL
}
\description{
Get Result from LipidTrendSE
}
\examples{
data("lipid_se_CL")
res_se <- analyzeLipidRegion(
    lipid_se=lipid_se_CL, ref_group="sgCtrl", split_chain=FALSE,
    chain_col=NULL, radius=3, own_contri=0.5, permute_time=100)
# Get complete result
results <- getResult(res_se)
}
