% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_info}
\alias{group_info}
\title{Example group information table for LipidTrend analysis}
\format{
A \code{data.frame} object of sample name, lable name, and group name
over 6 samples
}
\source{
Tomoyuki Shiota et al. ,Hepatoviruses promote very-long-chain fatty
acid and sphingolipid synthesis for viral RNA replication and quasi-enveloped
virus release. Sci Adv. 9(42):eadj4198
\url{https://www.science.org/doi/10.1126/sciadv.adj4198}.
}
\usage{
data(group_info)
}
\description{
Example group information table for LipidTrend analysis
}
\examples{
data(group_info)
}
