% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_se_2D}
\alias{lipid_se_2D}
\title{Example Dataset for two-dimensional data}
\format{
A \code{SummarizedExperiment} object with the following slots:
\describe{
\item{\code{colData}}{ A data frame with 6 observations on the following
3 variables, containing sample name, label name, and group name}
\item{\code{assay}}{ A 137*6 matrix containing lipid abundance data}
\item{\code{rowData}}{ A data frame with 137 observations on the following
2 variables containing total chain length and total double bond
characteristic information}
}
}
\usage{
data(lipid_se_2D)
}
\description{
Example Dataset for two-dimensional data
}
\examples{
data(lipid_se_2D)
}
