% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_se_CL}
\alias{lipid_se_CL}
\title{Example Dataset for one-dimensional data}
\format{
A \code{SummarizedExperiment} object with the following slots:
\describe{
\item{\code{colData}}{ A data frame with 6 observations on the following
3 variables, containing sample name, label name, and group name}
\item{\code{assay}}{ A 29*6 matrix containing lipid abundance data}
\item{\code{rowData}}{ A data frame with 29 observations on the following
1 variables containing chain characteristic information}
}
}
\usage{
data(lipid_se_CL)
}
\description{
Example Dataset for one-dimensional data
}
\examples{
data(lipid_se_CL)
}
