% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{LipidTrendSE-validity}
\alias{LipidTrendSE-validity}
\title{Validate LipidTrendSE objects}
\arguments{
\item{object}{A LipidTrendSE object to validate}
}
\value{
TRUE if valid, otherwise an error message
}
\description{
Validate LipidTrendSE objects
}
