% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{show,LipidTrendSE-method}
\alias{show,LipidTrendSE-method}
\title{Show method for LipidTrendSE objects}
\usage{
\S4method{show}{LipidTrendSE}(object)
}
\arguments{
\item{object}{A LipidTrendSE object}
}
\value{
LipidTrendSE object information
}
\description{
Show method for LipidTrendSE objects
}
