% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance_CL}
\alias{abundance_CL}
\title{Example lipid abundance data for 1 dimension LipidTrend analysis}
\format{
A \code{matrix} object of lipid abundance with 29 lipids over 6 samples
}
\source{
Tomoyuki Shiota et al. ,Hepatoviruses promote very-long-chain fatty
acid and sphingolipid synthesis for viral RNA replication and quasi-enveloped
virus release. Sci Adv. 9(42):eadj4198
\url{https://www.science.org/doi/10.1126/sciadv.adj4198}.
}
\usage{
data(abundance_CL)
}
\description{
Example lipid abundance data for 1 dimension LipidTrend analysis
}
\examples{
data(abundance_CL)
}
\keyword{datasets}
