% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeLipidRegion.R
\name{analyzeLipidRegion}
\alias{analyzeLipidRegion}
\title{Conduct statistical to analyze lipid features tendencies}
\usage{
analyzeLipidRegion(
  lipid_se,
  ref_group,
  split_chain = FALSE,
  chain_col = NULL,
  radius = 3,
  own_contri = 0.5,
  test = "t.test",
  abund_weight = TRUE,
  permute_time = 1e+05
)
}
\arguments{
\item{lipid_se}{A SummarizedExperiment object. Must contain following data:
\describe{
\item{Assay}{A matrix containing lipid abundance data, where rows
represent lipids and columns represent samples.}
\item{RowData}{A data frame of lipid features (e.g., double bond count,
chain length), with rows as lipids and columns as lipid features (
limited to 1 or 2 columns). The order of lipids must match the abundance
data. If RowData contains one column, a one-dimensional analysis will
be performed. If RowData includes two columns, a two-dimensional
analysis will be conducted.}
\item{ColData}{A data frame containing group information, where rows
represent sample names and columns must include sample name, label name,
and group, arranged accordingly.}
}}

\item{ref_group}{Character. Group name of the reference group. It must be
one of the group names in the colData group column.}

\item{split_chain}{Logical. If TRUE the results will split to shown by odd
and even chain. Default is \code{FALSE}.}

\item{chain_col}{Character. The column name of chain length. Set to NULL if
split_chain is FALSE. Default is \code{'NULL'}.}

\item{radius}{Numeric. Distance of neighbor to be included.
Default is \code{3}.}

\item{own_contri}{Numeric. Proportion of own contribution.
Default is \code{0.5}. To not over-emphasize the neighbor information,
we suggest to choose a value from 0.5 to 1.}

\item{test}{String. Choose statistic test from "t.test" and "Wilcoxon".
Default is \code{'t.test'}.}

\item{abund_weight}{Logical. Consider the average abundance as the weight in
the test statistic. Default is \code{TRUE}.}

\item{permute_time}{Integer. Permutation times. Default is \code{100000}.}
}
\value{
A LipidTrendSE object containing lipidomic feature testing result.
}
\description{
Calculate p-value of lipidomic features by permutation test.
The test of region statistics smoothing with Gaussian kernel integrates
neighbor's information and provides a stable testing result under small
sample size.
}
\examples{
data("lipid_se_CL")
res_se <- analyzeLipidRegion(
    lipid_se=lipid_se_CL, ref_group="sgCtrl", split_chain=FALSE,
    chain_col=NULL, radius=3, own_contri=0.5, permute_time=100)
}
\seealso{
\code{\link{plotRegion1D}} for one-dimensional visualization
\code{\link{plotRegion2D}} for two-dimensional visualization
}
