% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion2D.R
\name{plotRegion2D}
\alias{plotRegion2D}
\alias{plotRegion2D,LipidTrendSE-method}
\title{Plot region trends for two-dimensional lipid features}
\usage{
plotRegion2D(object, p_cutoff = 0.05, log2FC_cutoff = 3)

\S4method{plotRegion2D}{LipidTrendSE}(object, p_cutoff = 0.05, log2FC_cutoff = 3)
}
\arguments{
\item{object}{A LipidTrendSE object containing analysis results.}

\item{p_cutoff}{Numeric. Significance level. The range is 0 to 1.
Default is \code{0.05}.}

\item{log2FC_cutoff}{Numeric. Significant level of log2-transformed fold
change. Default is \code{3}.}
}
\value{
For split chain analyses, a list with two elements:
\item{even_result}{Heatmap for even chain features, or NULL if none exist}
\item{odd_result}{Heatmap for odd chain features, or NULL if none exist}
For non-split analyses, a single heatmap.
}
\description{
plotRegion2D visualizes lipid trend analysis results in two
dimensions. The returned heatmap will show the results.
}
\examples{
data("lipid_se_2D")
res_se <- analyzeLipidRegion(
    lipid_se=lipid_se_2D, ref_group="sgCtrl", split_chain=FALSE,
    chain_col=NULL, radius=3, own_contri=0.5, permute_time=100,
    abun_weight=TRUE)
plot_2D <- plotRegion2D(res_se, p_cutoff=0.05)
}
\seealso{
\code{\link{analyzeLipidRegion}} for generating the input LipidTrendSE object
}
